+++ Tests/Editor/Unity.Build.Windows.Classic.Tests/ClassicWindowsNonIncrementalPipelineTests.cs
@@ -1,36 +1,37 @@
 using NUnit.Framework;
 using UnityEditor;
+using Unity.Build.Classic.Private;
+using Unity.Build.Classic.Private.MissingPipelines;
 
 namespace Unity.Build.Windows.Classic.Tests
 {
     class ClassicWindowsNonIncrementalPipelineTests
     {
-#if UNITY_STANDALONE_WIN
-        const string k_PackagePath = "Packages/com.unity.platforms.windows/";
-        const string k_TestPath = "Tests/Editor/Unity.Build.Windows.Classic.Tests/";
-        const string k_BuildConfigurationPath = k_PackagePath + k_TestPath + "WindowsClassicBuildConfiguration.buildconfiguration";
-
-        [SetUp]
-        public void SetUp()
+        [Test]
+        public void TestBuildPipelineSelectorForWin64()
         {
-            BuildArtifacts.Clean();
-        }
+            var winPlatform = Platform.Windows;
+            var missingWindowsPlatform = new MissingPlatform(winPlatform.Name);
 
-        [TearDown]
-        public void TearDown()
-        {
-            BuildArtifacts.Clean();
+            var selector = new BuildPipelineSelector();
+            var pipeline = selector.SelectFor(winPlatform);
+            Assert.AreEqual(pipeline.GetType(), typeof(WindowsClassicNonIncrementalPipeline));
+
+            // Since Windows 64 pipeline is implemented in com.unity.platforms.windows
+            // Selector will return null pipeline instead of MissingNonIncrementalPipeline
+            pipeline = selector.SelectFor(missingWindowsPlatform);
+            Assert.AreEqual(pipeline, null);
         }
 
         [Test]
-        public void Build()
+        public void TestBuildPipelineSelectorForWin32()
         {
-            var config = AssetDatabase.LoadAssetAtPath<BuildConfiguration>(k_BuildConfigurationPath);
-            var result = config.Build();
-            if (result.Failed)
-                result.LogResult();
-            Assert.That(result.Succeeded, Is.True);
+            var missingWindows32Platform = new MissingPlatform("Windows32");
+
+            var selector = new BuildPipelineSelector();
+            var pipeline = selector.SelectFor(missingWindows32Platform);
+            // There's no pipeline implemented for Windows 32 bit, thus MissingNonIncrementalPipeline will be used
+            Assert.AreEqual(pipeline.GetType(), typeof(MissingNonIncrementalPipeline));
         }
-#endif
     }
 }